OPIK-610 move retry logic to ChatCompletionService
